flops_manager_package/pyproject.toml
[tool.poetry]
name = "flops_manager"
version = "0.1.0"
description = "The FLOps Manager is the main control component for the FLOps plugin of Oakestra"
authors = ["Alexander Malyuk <malyuk.alexander1999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^2.0"
flask = "^3.0.2"
flask-openapi3 = "^3.0.2"
requests = "^2.28.1"
docker = "^7.0.0"
gitpython = "^3.1.42"
PyGithub = "^2.2.0"
pymongo = "^4.6.3"
pydantic = "^2.6.4"
waitress = "3.0.0"
icecream = "^2.1.3"
flops-utils = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
oak = "flops_manager.main:main"

[tool.black]
line-length = 100
